Create MinesFactory and OctoCheckService - Mines factory takes the creating of the mines out of NewMineField. - OctoCheckService gets the 8 squares around the index square.
